requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
